ae_debug/requirements.txt
torch
torchvision
tqdm
pyyaml
matplotlib
Pillow
scikit-learn 
torch-fidelity 